fixes duplicated keys and fixes links to presons
