+++ src/common/Credits/index.js
@@ -1,5 +1,6 @@
 import {
   FullName,
+  PersonLink,
   Portrait,
   PortraitBackground,
   Role,
@@ -7,20 +8,25 @@ import {
   Tile,
 } from "./styled";
 
-const Credits = ({ path, name, role }) => (
-  <Tile>
-    <PortraitBackground>
-      {path ? (
-        <Portrait src={`https://image.tmdb.org/t/p/w500/${path}`} />
-      ) : (
-        <Portrait />
-      )}
-    </PortraitBackground>
-    <Storage>
-      <FullName>{name}</FullName>
-      <Role>{role}</Role>
-    </Storage>
-  </Tile>
+const Credits = ({ path, name, role, id }) => (
+  <PersonLink
+    key={name}
+    to={`/people/${id}`}
+  >
+    <Tile>
+      <PortraitBackground>
+        {path ? (
+          <Portrait src={`https://image.tmdb.org/t/p/w500/${path}`} />
+        ) : (
+          <Portrait />
+        )}
+      </PortraitBackground>
+      <Storage>
+        <FullName>{name}</FullName>
+        <Role>{role}</Role>
+      </Storage>
+    </Tile>
+  </PersonLink>
 );
 
 export default Credits;
