--- src/features/peopleList/PeoplePage/People/styled.js
@@ -1,8 +1,0 @@
-import { Link } from "react-router-dom";
-import styled from "styled-components";
-
-export const PersonLink = styled(Link)`
-display: flex;
-height: 100%;
-text-decoration: none;
-`
